+++ backend/lib/b3_wallet_lib/src/ledger/btc/utils.rs
@@ -5,6 +5,20 @@ pub const MAX_ENCODED_SIGNATURE_LEN: usize = 73;
 
 pub const PUBKEY_LEN: usize = 32;
 
+/// Computes an estimate for the size of transaction (in vbytes) with the given number of inputs and outputs.
+pub fn tx_vsize_estimate(input_count: u64, output_count: u64) -> u64 {
+    // See
+    // https://github.com/bitcoin/bips/blob/master/bip-0141.mediawiki
+    // for the transaction structure and
+    // https://bitcoin.stackexchange.com/questions/92587/calculate-transaction-fee-for-external-addresses-which-doesnt-belong-to-my-loca/92600#92600
+    // for transaction size estimate.
+    const INPUT_SIZE_VBYTES: u64 = 68;
+    const OUTPUT_SIZE_VBYTES: u64 = 31;
+    const TX_OVERHEAD_VBYTES: u64 = 11;
+
+    input_count * INPUT_SIZE_VBYTES + output_count * OUTPUT_SIZE_VBYTES + TX_OVERHEAD_VBYTES
+}
+
 const MOCK_SIG: [u8; MAX_ENCODED_SIGNATURE_LEN] = [
     0x30, 70, 0x02, 33, 0x00, 0x8f, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
     0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
